dataset/requirements.txt
python-dotenv
datasets
huggingface_hub
pandas
mysql-connector-python
